version lock

+++ mounts/modules/install.sh
@@ -1,21 +1,37 @@
+# Function to checkout a git version
+version_checkout() {
+    # Check if the version is specified
+    if [ ! -z "$1" ]; then
+        # Checkout the version
+        (cd "$(dirname "$0")/$2" && git checkout $1)
+    fi
+}
+
 # Load the modules file to get the git urls
 while read -r module; do
+    # Skip empty lines and comments
+    if [ -z "$module" ] || [[ "$module" =~ ^# ]] ; then
+        continue
+    fi
     echo $module
     # Split by space and get the first element
     url=$(echo $module | awk '{print $1}')
     name=$(echo $module | awk '{print $2}')
+    version=$(echo $module | awk '{print $3}')
     # if module already exists.
     if [ -d "$(dirname "$0")/$name" ]; then
         # check if the module is the right git repo
         if [ "$(cd $(dirname "$0")/$name && git remote get-url origin)" != "$url" ]; then
             rm -rf "$(dirname "$0")/$name"
         else
+            version_checkout $version $name
             # Pull the latest changes
             (cd "$(dirname "$0")/$name" && git pull)
         fi
     else 
         # Clone the module into the modules directory
-        (cd "$(dirname "$0")" && git clone $module)
+        (cd "$(dirname "$0")" && git clone $url $name)
+        version_checkout $version $name
     fi
 done < $(dirname "$0")/modules
